styling fixes
