+++ .gitignore
@@ -9,3 +9,4 @@ npm-debug.*
 *.orig.*
 web-build/
 web-report/
+*.sh
